Refactorización a estructura MVC
